requirements.txt
# Core Dependencies
torch==2.8.0
torchvision==0.23.0
numpy==2.0.2
Pillow==11.3.0

# Web Framework
Flask==3.1.2
Werkzeug==3.1.3

# Data Science
pandas==2.3.3
matplotlib==3.9.4
scikit-learn==1.6.1
scikit-plot==0.3.7
scipy==1.13.1

# Jupyter (Optional - for notebooks)
jupyterlab==4.5.0
ipykernel==6.31.0

# Utilities
tqdm
